+++ pandas_alive/__init__.py
@@ -26,7 +26,7 @@ version = "0.2.3"
 
 plot_animated = CachedAccessor("plot_animated", AnimatedAccessor)
 pd.DataFrame.plot_animated = plot_animated
-pd.Series.plot_animated = plot
+pd.Series.plot_animated = plot_animated
 
 # Define plot_animated method for GeoPandas and Series: